make gemini download more reliable - "invalid url"
while testing if SNI is working on gemini://drewdevault.com I got tons of errors
stating "invalid url". I do not know why this change works, but before it was
working only randomly, now it seems to work all the time. Maybe this is also a
server side issue, but a workaround is nice nonetheless.