some more code quality improvements
Made not having a specific view an error so the respective functions in Layout
do not return an Option, but the real content. If a view is missing, this
should be considered an unrecoverable error.
Added some more uses of unwrap_or instead of if's.
Simplified how the current view is retrieved everywhere I found it.