change known hosts format and canonicalize hosts
This resolves some issues I had when a certificate changed.
The problem was due to missing normalization of the host names before
writing them into the known hosts file.
This is a soft breaking change, i.e. the old format is still understood.

This change is opposed to the doc-comment of Certificate (removed)
because for TOML there is no difference between nested "Tables" and an
array, so it should equally well be possible to adapt to future changes
of ip-adresses or cipher types. At the same time it is much easier to
serialize the known hosts this way.