remove default URL scheme detection/insertion
Because of the about:* URLs, the detection does not work by looking for "://"
and because all URLs may contain port segments, does also not work by looking
for just ":". Not even the url crate could detect that "jan.bio:70" is
supposed to be another URL and says that "jan.bio" is the scheme part.